+++ src/apps/guanabana/factory.ts
@@ -2,7 +2,7 @@ import { FactoryOlula } from "../olula/factory.ts";
 import { FactoryVentasGUA } from "./contextos/ventas/factory.ts";
 
 export class FactoryGUA extends FactoryOlula {
-    static Ventas = FactoryVentasGUA
+    Ventas = FactoryVentasGUA;
 }
 
 export default FactoryGUA;
