+++ src/apps/olula/factory.ts
@@ -1,7 +1,10 @@
 import { FactoryVentasOlula } from '../../contextos/ventas/factory.ts';
+import { otrosMenu } from './tmp_menu.ts';
 
 export class FactoryOlula {
-    static Ventas = FactoryVentasOlula
+    Inicio = { menu: { nombre: "Inicio", url: "/", icono: "inicio" } };
+    Ventas = FactoryVentasOlula;
+    Otros = { menu: otrosMenu };
 }
 
 export default FactoryOlula;
